+++ packages/hpe-design-tokens/src/semantic_color_normalization_core.ts
@@ -1,0 +1,221 @@
+import {
+  SEMANTIC_COLOR_NORMALIZATION_SEGMENTS,
+  SEMANTIC_COLOR_ROLE_VARIANTS,
+  SEMANTIC_COLOR_SCALES,
+  SEMANTIC_COLOR_STATES,
+  type SemanticColorScale,
+  type SemanticColorState,
+  type SemanticColorTarget,
+} from './semantic_color_core.js';
+
+export const SEMANTIC_COLOR_EXPORT_EXCEPTION_ALIASES = [
+  'color/focus/support',
+  'color/transparent',
+] as const;
+
+export const SEMANTIC_COLOR_IMPORT_EXCEPTION_SECTIONS = [
+  'color/focus',
+] as const;
+
+const semanticInteractions = SEMANTIC_COLOR_STATES.filter(
+  state => state !== 'REST',
+);
+const semanticProminences = SEMANTIC_COLOR_SCALES;
+
+const semanticInteractionSet = new Set<string>(semanticInteractions);
+const semanticProminenceSet = new Set<string>(semanticProminences);
+
+export const normalizationSegmentSet = new Set<string>(
+  SEMANTIC_COLOR_NORMALIZATION_SEGMENTS,
+);
+
+function isCanonicalState(value: string): value is SemanticColorState {
+  return SEMANTIC_COLOR_STATES.includes(value as SemanticColorState);
+}
+
+function isCanonicalScale(value: string): value is SemanticColorScale {
+  return SEMANTIC_COLOR_SCALES.includes(value as SemanticColorScale);
+}
+
+export function canonicalTokenPathSegments(input: string | string[]) {
+  const raw = Array.isArray(input) ? input.join('/') : input;
+  const cleaned = raw
+    .trim()
+    .replace(/[{}]/g, '')
+    .replace(/\./g, '/')
+    .replace(/\/+/g, '/')
+    .replace(/^\//, '')
+    .replace(/\/$/, '');
+
+  const segments = cleaned.split('/').filter(Boolean);
+
+  if (segments[0] === 'hpe') {
+    return segments.slice(1);
+  }
+
+  return segments;
+}
+
+export function expandCompactRoleSegment(
+  target: SemanticColorTarget,
+  segment: string,
+  knownFamily?: string,
+) {
+  const partTokens = segment.split('-').filter(Boolean);
+  if (partTokens.length <= 1) {
+    return [segment];
+  }
+
+  if (!knownFamily) {
+    const variantFamilies = SEMANTIC_COLOR_ROLE_VARIANTS[
+      target as keyof typeof SEMANTIC_COLOR_ROLE_VARIANTS
+    ] as Record<string, readonly string[]> | undefined;
+    const compactFamily = partTokens[0];
+    const compactIntent = partTokens.slice(1).join('-');
+    const compactIntentOptions = variantFamilies?.[compactFamily];
+
+    if (compactIntentOptions?.includes(compactIntent)) {
+      return [compactFamily, compactIntent];
+    }
+  }
+
+  const interactionCandidate = partTokens[partTokens.length - 1];
+  const hasInteraction = isCanonicalState(
+    interactionCandidate as SemanticColorState,
+  );
+  const scaleIndex = hasInteraction
+    ? partTokens.length - 2
+    : partTokens.length - 1;
+  const scaleCandidate = partTokens[scaleIndex];
+  const normalizedScale = scaleCandidate?.toLowerCase();
+  const hasScale =
+    !!normalizedScale &&
+    (isCanonicalScale(normalizedScale as SemanticColorScale) ||
+      scaleCandidate === 'DEFAULT');
+
+  if (!hasScale && !hasInteraction) {
+    return [segment];
+  }
+
+  let intentEndIndex = scaleIndex;
+  if (hasInteraction && !hasScale) {
+    intentEndIndex = partTokens.length - 1;
+  }
+  const intentParts = partTokens.slice(0, intentEndIndex);
+  if (intentParts.length === 0) {
+    return [segment];
+  }
+
+  const expanded = [intentParts.join('-')];
+  if (hasScale) {
+    expanded.push(scaleCandidate);
+  }
+  if (hasInteraction) {
+    expanded.push(interactionCandidate);
+  }
+
+  return expanded;
+}
+
+export function normalizeRoleSegments(
+  target: SemanticColorTarget,
+  segments: string[],
+) {
+  if (segments.length === 1) {
+    return expandCompactRoleSegment(target, segments[0]);
+  }
+
+  if (segments.length === 2 && segments[1].includes('-')) {
+    return [
+      segments[0],
+      ...expandCompactRoleSegment(target, segments[1], segments[0]),
+    ];
+  }
+
+  return segments;
+}
+
+export function colorNameToHierarchyPartsCore(
+  colorVariableName: string,
+): string[] {
+  const parts = colorVariableName.split('/');
+
+  if (parts[0] !== 'color') {
+    return parts;
+  }
+
+  if (parts[1] === 'background' && parts[2] === 'accent') {
+    const section = parts.slice(0, 3);
+    const tail = parts.slice(3).join('-');
+    if (!tail) {
+      return section;
+    }
+
+    const expanded = expandCompactRoleSegment('background', tail, 'accent');
+
+    if (expanded.length === 1 && expanded[0] === tail) {
+      return [...section, tail];
+    }
+
+    return [...section, ...expanded];
+  }
+
+  return colorVariableName.replaceAll('-', '/').split('/');
+}
+
+export function normalizeColorVariableNameFromFigmaCore(
+  colorVariableName: string,
+): string {
+  const temp = colorNameToHierarchyPartsCore(colorVariableName);
+  if (
+    !(SEMANTIC_COLOR_EXPORT_EXCEPTION_ALIASES as readonly string[]).includes(
+      temp.join('/'),
+    )
+  ) {
+    if (
+      !semanticInteractionSet.has(temp[temp.length - 1]) ||
+      temp.join('/') === 'color/focus'
+    ) {
+      temp.push('REST');
+    }
+
+    if (!semanticProminenceSet.has(temp[temp.length - 2])) {
+      temp.splice(temp.length - 1, 0, 'DEFAULT');
+    }
+  }
+
+  return temp.join('/');
+}
+
+export function tokenAliasToFigmaAliasCore(alias: string): string {
+  const isColor = /^color/.test(alias);
+  let adjustedName = alias;
+
+  if (isColor) {
+    let parts = adjustedName.split('/');
+    parts = parts.filter(part => !normalizationSegmentSet.has(part));
+    let section = parts.slice(0, 2).join('/');
+    let name = parts.slice(2).join('-');
+
+    const isAccentBackground =
+      section === 'color/background' && parts[2] === 'accent';
+
+    if (isAccentBackground) {
+      section = parts.slice(0, 3).join('/');
+      name = parts.slice(3).join('-');
+    }
+
+    if (
+      (SEMANTIC_COLOR_IMPORT_EXCEPTION_SECTIONS as readonly string[]).includes(
+        section,
+      )
+    ) {
+      section = parts.slice(0, 1).join('/');
+      name = parts.slice(1).join('-');
+    }
+
+    adjustedName = `${section}${name ? `/${name}` : ''}`;
+  }
+
+  return adjustedName;
+}
