+++ packages/hpe-design-tokens/src/semantic_color_parser.ts
@@ -1,7 +1,6 @@
 // Parses semantic color token paths into canonical metadata and exports
 // helpers to serialize that metadata for build artifacts.
 import {
-  SEMANTIC_COLOR_NORMALIZATION_SEGMENTS,
   SEMANTIC_COLOR_ROLE_FAMILIES_BY_TARGET,
   SEMANTIC_COLOR_ROLE_VARIANTS,
   SEMANTIC_COLOR_SCALES,
@@ -13,6 +12,11 @@ import {
   type SemanticColorTokenMetadata,
   type SemanticColorTokenMetadataMap,
 } from './semantic_color_core.js';
+import {
+  canonicalTokenPathSegments,
+  normalizationSegmentSet,
+  normalizeRoleSegments,
+} from './semantic_color_normalization_core.js';
 
 export type SemanticColorTokenParseErrorCode =
   | 'NOT_A_COLOR_TOKEN'
@@ -69,25 +73,6 @@ export type SemanticColorMetadataExportOptions =
 
 type TokenTree = Record<string, unknown>;
 
-function canonicalTokenPathSegments(input: string | string[]) {
-  const raw = Array.isArray(input) ? input.join('/') : input;
-  const cleaned = raw
-    .trim()
-    .replace(/[{}]/g, '')
-    .replace(/\./g, '/')
-    .replace(/\/+/g, '/')
-    .replace(/^\//, '')
-    .replace(/\/$/, '');
-
-  const segments = cleaned.split('/').filter(Boolean);
-
-  if (segments[0] === 'hpe') {
-    return segments.slice(1);
-  }
-
-  return segments;
-}
-
 function isCanonicalTarget(value: string): value is SemanticColorTarget {
   return SEMANTIC_COLOR_TARGETS.includes(value as SemanticColorTarget);
 }
@@ -98,86 +83,6 @@ function isCanonicalState(value: string): value is SemanticColorState {
 
 function isCanonicalScale(value: string): value is SemanticColorScale {
   return SEMANTIC_COLOR_SCALES.includes(value as SemanticColorScale);
-}
-
-function expandCompactRoleSegment(
-  target: SemanticColorTarget,
-  segment: string,
-  knownFamily?: string,
-) {
-  const partTokens = segment.split('-').filter(Boolean);
-  if (partTokens.length <= 1) {
-    return [segment];
-  }
-
-  if (!knownFamily) {
-    const variantFamilies =
-      SEMANTIC_COLOR_ROLE_VARIANTS[
-        target as keyof typeof SEMANTIC_COLOR_ROLE_VARIANTS
-      ] as Record<string, readonly string[]> | undefined;
-    const compactFamily = partTokens[0];
-    const compactIntent = partTokens.slice(1).join('-');
-    const compactIntentOptions = variantFamilies?.[compactFamily];
-
-    if (compactIntentOptions?.includes(compactIntent)) {
-      return [compactFamily, compactIntent];
-    }
-  }
-
-  const interactionCandidate = partTokens[partTokens.length - 1];
-  const hasInteraction = isCanonicalState(
-    interactionCandidate as SemanticColorState,
-  );
-  const scaleIndex = hasInteraction
-    ? partTokens.length - 2
-    : partTokens.length - 1;
-  const scaleCandidate = partTokens[scaleIndex];
-  const normalizedScale = scaleCandidate?.toLowerCase();
-  const hasScale =
-    !!normalizedScale &&
-    (isCanonicalScale(normalizedScale as SemanticColorScale) ||
-      scaleCandidate === 'DEFAULT');
-
-  if (!hasScale && !hasInteraction) {
-    return [segment];
-  }
-
-  let intentEndIndex = scaleIndex;
-  if (hasInteraction && !hasScale) {
-    intentEndIndex = partTokens.length - 1;
-  }
-  const intentParts = partTokens.slice(0, intentEndIndex);
-  if (intentParts.length === 0) {
-    return [segment];
-  }
-
-  const expanded = [intentParts.join('-')];
-  if (hasScale) {
-    expanded.push(scaleCandidate);
-  }
-  if (hasInteraction) {
-    expanded.push(interactionCandidate);
-  }
-
-  return expanded;
-}
-
-function normalizeRoleSegments(
-  target: SemanticColorTarget,
-  segments: string[],
-) {
-  if (segments.length === 1) {
-    return expandCompactRoleSegment(target, segments[0]);
-  }
-
-  if (segments.length === 2 && segments[1].includes('-')) {
-    return [
-      segments[0],
-      ...expandCompactRoleSegment(target, segments[1], segments[0]),
-    ];
-  }
-
-  return segments;
 }
 
 export function parseSemanticColorTokenMetadata(
@@ -257,10 +162,7 @@ export function parseSemanticColorTokenMetadata(
     }
   }
 
-  const normalizationSegments = new Set<string>(
-    SEMANTIC_COLOR_NORMALIZATION_SEGMENTS,
-  );
-  const roleParts = rest.filter(part => !normalizationSegments.has(part));
+  const roleParts = rest.filter(part => !normalizationSegmentSet.has(part));
 
   if (roleParts.length === 0) {
     return {
